pyproject.toml
[project]
name = "playlet-clip"
version = "2.0.0"
description = "AI-powered short drama auto-editing tool with ASR, TTS and Gradio UI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "anning", email = "anningforchina@gmail.com" }
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",

    # ASR
    "funasr>=1.0",
    "modelscope>=1.10",

    # TTS - CosyVoice dependencies (optional, install from source)
    "torch>=2.0",
    "torchaudio>=2.0",
    "transformers>=4.35",
    "onnxruntime>=1.16",

    # TTS - Edge-TTS fallback (lightweight, works out of box)
    "edge-tts>=6.1",

    # LLM
    "openai>=1.0",

    # Video
    "moviepy>=1.0",
    "ffmpeg-python>=0.2",

    # Audio
    "mutagen>=1.47",
    "librosa>=0.10",
    "soundfile>=0.12",

    # UI
    "gradio>=4.0",

    # Async
    "aiofiles>=23.0",
    "httpx>=0.25",

    # Logging
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
playlet-clip = "playlet_clip.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/playlet_clip"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
